refactor: 💡 route control
